update: rename directory to i18n + rename variables
